moves reset progress and overlay code into MainComponent
